Emacs: open new tab in same buffer as current
Emacs has various project-level commands that operate on the
current buffer. When creating an elscreen tab I'd prefer to
stay within the context of the project.

Ideally by default the projectile commands would act relative
to the directory of the PWD of my parent terminal session,
but instead it changes to match the file.